Move AsyncTaskResult to c.a.f.asynchronous.asynctasks package
For more general purpose.
