+++ app/src/main/java/com/amaze/filemanager/ui/dialogs/SftpConnectDialog.java
@@ -46,7 +46,7 @@ import com.amaze.filemanager.database.UtilsHandler;
 import com.amaze.filemanager.fragments.MainFragment;
 import com.amaze.filemanager.filesystem.ssh.SshClientUtils;
 import com.amaze.filemanager.filesystem.ssh.SshConnectionPool;
-import com.amaze.filemanager.asynchronous.asynctasks.ssh.AsyncTaskResult;
+import com.amaze.filemanager.asynchronous.asynctasks.AsyncTaskResult;
 import com.amaze.filemanager.asynchronous.asynctasks.ssh.PemToKeyPairTask;
 import com.amaze.filemanager.asynchronous.asynctasks.ssh.SshAuthenticationTask;
 import com.amaze.filemanager.asynchronous.asynctasks.ssh.GetSshHostFingerprintTask;
